Split methods into separate files
The mono file was getting too big.
The methods were split into these logical units:
- Date
- DateTime
- Day
- Time
