+++ simpledate/src/main/kotlin/com/parassidhu/simpledate/Day.kt
@@ -1,0 +1,8 @@
+package com.parassidhu.simpledate
+
+import java.util.*
+
+fun Date?.toDay(): String {
+  val pattern = "EEEE"
+  return dateAsString(this, pattern)
+}
